requirements.txt
gradio
openai
pandas
selenium
webdriver-manager
python-dotenv